requirements.txt
azure-functions
tweepy
langchain
openai
tiktoken
unstructured